ola/backend/requirements.txt
amqp==2.3.2
anyjson==0.3.3
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
billiard==3.5.0.4
celery==4.2.1
decorator==4.3.0
Django==1.11.14
django-cors-headers==2.3.0
djangorestframework==3.8.2
enum34==1.1.6
ipython==5.7.0
ipython-genutils==0.2.0
kombu==4.2.1
MySQL-python==1.2.3
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.2.0
pytz==2018.5
redis==2.10.6
scandir==1.7
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
vine==1.1.4
wcwidth==0.1.7
